+++ src/models/index.ts
@@ -1,6 +1,7 @@
 import { fail } from '@noeldemartin/utils';
 
 import { SoukaiError } from '@/errors';
+import ModelKey from '@/models/ModelKey';
 import type { Model } from '@/models/Model';
 import type { ModelConstructor } from '@/models/inference';
 
@@ -29,6 +30,8 @@ export {
 } from './fields';
 
 export type { Attributes, AttributeValue } from './attributes';
+
+export { ModelKey };
 
 const bootedModels: Map<string, typeof Model> = new Map;